requirements.txt
yq
rich
python-dotenv